Clean up - still running into issues with multiple publish cards
